Add the concept of a page with elements .. with elements ... with elements.
The basic idea is that an HTML page typically has elements (bits of HTML) on it
that have nothing to do with the actual page's true content and purpose;
they're just part of the site furniture. Having to handle these elements on a
per-page basis is extremely tedious and error-prone.

Elements are also a convenient way of encapsulating functionality that a page
includes, i.e. a list of recent news posts, an authentication component, etc.
By including an element identifer (not currently implemented) in the emitted
HTML form submissions, AJAX requests, etc can be directed to the element
itself.

For now only pages and their elements are supported. Elements with elements are
completely ignored!

This change has moved a relatively page-specific templating helper
(templating.template) to the page module (page.template). I suspect the
decorator is only really useful for rendering HTML pages and it now includes
some page-specific details, e.g. a tag to lookup a page element.

Finally (I think) I re-did the meta classes to make the inheritance clearer and
therefore more understandable as well as fixing the collection of annotated
methods so they don't splat the parent classes' annotations.