Use decoded path segments when locating resources.
By using a URL instance the path segments are decoded from their URL
encoding (as provided by WebOb) to proper strings.

I also decided that Request.path_segments, that was actually returning
the path segments relative to the application, had no right to exist.
The application now handles the application-relativeness of the URL.