Remove all dependence on Pylons and specific templating engines.
Restish now has no support for specific templating engines (it only
supported Mako anyway ;-)). Instead the developer is expected to
configure the restish.templating.renderer environ key with a callable
that the templating module can use.

Also, we were using webhelper's literal to tell the rendering engine not
to escape fragments of HTML. literal only works with certain templating
engines and, even then, only those that can be configured to use
webhelper's escape function.

I've removed literal. It's now up to the template that the fragment is
included in to ensure the HTML makes it to the output stream unescaped.

All templating engines (that I know of) either have a way to tell the
engine not to escape something when the engine is escaping everything
automatically.