Allow resource_child(request, segments) to return just a resource.
If the result of calling resource_child(...) is not a tuple then it is
considered to be the next resource to lookup children on and the list of
remaining segments is left unchanged.

As a bonus this allows resource_child() to return just None to signify a
not found condition. Test added for this so considered official
behaviour from now.