Guard checkers should now raise an error and the overall response to failure can be customised.
A checker is now expected to raise a GuardError on failure, containing a
descriptive error message. The error message is added to the "401
Unauthorized" response.

guard has also been extended so its error handling can be customised. By
default, guard raises an http.UnauthorizedError but applications can
change that behaviour. For instance, an application may create a pretty,
descriptive page that includes a login form instead.