Add a lib.templating module to the paster templates.
The addition of this makes a restish project practically ready to go for
HTML templating "out of the box". Uncomment your preferred templating
engine in the templating module and start using it in resources. Tada!

This doesn't force a templating engine on anyone, and it's a no-brainer
to remove it completely anyway, but it's worth it being there by
default.