Enhance @child decorators and simple resource_child "overriding".
@child decorators now use a pattern matching algorithm instead of a
simple dict lookup.

The current behaviour - @child() and @child('segment') - is still
fully supported. However, you can now match deeper paths using
@child('assets/css') and extract path segments to args using
@child('users/{username}').

Any extracted args are passed to the decorated method as keyword args:

    @resource.child('users/{username}')
    def user(self, request, segments, username):
        ...

Matched segments are consumed as before so the 'segments' passed to a
@resource.child method are the segments of child resources.

Internally, a string passed to resource.child is converted into a
TemplateChildMatcher instance. The matching process calls those
instances to test for matches.

If something other than a string is passed to resource.child then it is
used as-is. The only requirement of a custom matcher is that it provides
a score attribute (to allow for specificity ordering) and is callable.

In addition, a "match anything", matcher is provided as resource.any.
This can be used to simplify overriding resource_child(request,
segments):

    @resource.child(resource.any)
    def any_child(self, request, segments):
        ...

It should now be extremely unusual to need to override
Resource.resource_child.