URLAccessor consistency and functionality improvements.
Various URLAccessor names were changed to match the names of
http.Request (and therefore webob.Request) attributes. Also added
URLAccessor attributes for the Request attributes added the other day.

The URLAccessor also include a new(str) method to create a new URL
instance.

Finally, the URLAccessor has been renamed in templates from 'url' to
'urls' to signal that it's not actually a URL but a collection of URLs.

Fixes issue #262.