Pass the request to @child matcher functions.
Thinking of a compelling usecase for passing the request to a @child
matcher was quite hard work (most of the time things were better solved
with a resource) but it's quite possible that some information the
@child matcher needs is in configuration inside the request.environ.

In any case, the request is utterly fundermental to just about
everything restish does that it seems wrong not to pass it around. It
basically represents and contains everything the application can do.

Closes issue #264.