Extend the signature of a "renderer" to include the output encoding and only encode page output by default.
Most templating engines don't like to be fed a string unless the string
is unicode-safe, i.e. unicode(some_text) does not raise an exception.

Page output is typically sent directly to a web client, but fragments of
HTML are often included in a page before being bundled up into the final
document. Therefore, only Rendering.page's decorator now specifies an
encoding; everything else renders to a unicode string.

It's up to the renderer to decide how best to encode or not encode. Some
engines will produce either stream given the correct information, some
return unicode whatever and the result must then be encoded.

Note: part of this change was to split Rendering.page's content_type
arg into two: type (the MIME type) and encoding (the charset). I believe
this makes things more flexible anyway - it's a bit of a pain having to
include the type and charset in a single string.