Added Request.application_path.
Templates and other code sometimes need the path the WSGI application is
mounted at in the WSGI server (not always the root of the server).

While the information was already available as
request.application_url.path it seems nice to shorten that for the
benefit of template writers and because we have path vs path_url
already.