Extend the @resource.GET et al decorators to support functions.
A function could already be used as a resource. However, the function
would always return the same response regardless of the request's
method or content negotiation -related headers.

The request method decorators can now be applied to functions, and not
just methods, to ensure that the request matches the requirements of the
function, including returning sensible HTTP error responses.

The following class and function are basically equivalent:

    class SimpleResource(resource.Resource):
        @resource.GET(accept='text/plain')
        def text(self, request):
            return http.ok(...)

    @resource.GET(accept='text/plain')
    def simple_resource(request)
        return http.ok(...)

whereas before this change you would need to add lots of request checks
to the simple_resource function to get the same behaviour.