Convert most of the tests to use WebTest instead of the wsgi_out util and hand-built requests. Deprecate wsgi_out.
WebTest has a nicer API but it also provides lint-like checks on the
response so it's actually checking the sanity of the response at the
same time as the actual tests.

The main reason for switching to WebTest is because there's some new
tests I need to add that highlight broken behaviour. I wasn't expecting
any failures to show up just yet. As it happens, the "method no allowed"
response was already wrong (fixed now).