Fix http.created signature.
'body' was a required arg but headers was optional. That makes no sense
- if there's a body there has to be a content type header too so headers
is no longer optional.

In addition (since I was breaking API anyway) I made reversed the order
of the body and headers args to be more consistent with the Response
initialiser, i.e. it's now create(location, headers, body).