Add a default HEAD implementation but allow Resource subclasses to provide a better HEAD handler when necessary.
The default HEAD implementation calls the resource as if the request was
a GET and then discards any content but retaining the content length
header.

A Resource subclass can provide a better implementation by simply
including a @HEAD method, e.g. to avoid loading a large blob of content
from a database if the size of the blob is known.