Add test for request handler methods that leaked to other classes in the hierarchy.
The problem was actually fixed in an earlier commit
(ace522e71824e0847e7eedfc031c7e86270c01de) but I only just noticed it
and want to make sure this never, ever happens again as it's a truly
horrible bug.

I also added a similar test for the resource's child factories although
it was actually ok.