Abort best dispatcher matching after no content type match.
It makes sense to abort early anyway, but it was also causing an error
(possibly a bug) when calling mimeparse to find `accept` matches.

No match because of the content type now returns a "415 Unsupported
Media Type" response. This is actually a small change in behaviour but
should have minimal impact since it's still an error and is now a more
accurate description of the error.